# openapi.json
{
  "components": {
    "schemas": {
...
      "ConsentPurpose": {
        "description": "Consent events captured at enrollment or upload (04 \u00a75.2).",
        "enum": [
          "data_processing",
          "document_upload",
          "sms_notifications",
          "email_notifications",
          "photo_publish"
        ],
        "title": "ConsentPurpose",
        "type": "string"
      },
      "ConsentRecordRequest": {
        "additionalProperties": false,
        "properties": {
          "granted": {
            "title": "Granted",
            "type": "boolean"
          },
          "purpose": {
            "$ref": "#/components/schemas/ConsentPurpose"
          },
          "text_version": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[^\\x00]*$",
            "title": "Text Version",
            "type": "string"
          }
        },
        "required": [
          "purpose",
          "granted",
          "text_version"
        ],
        "title": "ConsentRecordRequest",
        "type": "object"
      },
      "ConsentResponse": {
        "properties": {
          "granted": {
            "title": "Granted",
            "type": "boolean"
          },
          "granted_via": {
            "$ref": "#/components/schemas/GrantedVia"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "purpose": {
            "$ref": "#/components/schemas/ConsentPurpose"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          },
          "text_version": {
            "title": "Text Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "purpose",
          "granted",
          "granted_via",
          "text_version",
          "recorded_at"
        ],
        "title": "ConsentResponse",
        "type": "object"
      },
      "ConsentStateOut": {
        "description": "The current state of one purpose: the newest row, or never recorded.",
        "properties": {
          "granted": {
            "title": "Granted",
            "type": "boolean"
          },
          "granted_via": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GrantedVia"
              },
              {
                "type": "null"
              }
            ]
          },
          "purpose": {
            "$ref": "#/components/schemas/ConsentPurpose"
          },
          "recorded_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorded At"
          },
          "text_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Version"
          }
        },
        "required": [
          "purpose",
          "granted",
          "granted_via",
          "text_version",
          "recorded_at"
        ],
        "title": "ConsentStateOut",
        "type": "object"
      },
      "ConsentStateResponse": {
        "properties": {
          "consents": {
            "items": {
              "$ref": "#/components/schemas/ConsentStateOut"
            },
            "title": "Consents",
            "type": "array"
          },
          "farmer_id": {
            "title": "Farmer Id",
            "type": "integer"
          }
        },
        "required": [
          "farmer_id",
          "consents"
        ],
        "title": "ConsentStateResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "OpenAPI shape of the error envelope, so documented error responses match\nwhat the handler emits (keeps the Schemathesis contract test honest).",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "FarmerEnrollRequest": {
        "properties": {
          "address": {
            "anyOf": [
              {
                "pattern": "^[^\\x00]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "consent_granted": {
            "title": "Consent Granted",
            "type": "boolean"
          },
          "consent_text_version": {
            "pattern": "^[^\\x00]*$",
            "title": "Consent Text Version",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "pattern": "^[^\\x00]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "first_name": {
            "pattern": "^[^\\x00]*$",
            "title": "First Name",
            "type": "string"
          },
          "last_name": {
            "pattern": "^[^\\x00]*$",
            "title": "Last Name",
            "type": "string"
          },
          "number_of_hives": {
            "anyOf": [
              {
                "maximum": 2147483647.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number Of Hives"
          },
          "password": {
            "minLength": 1,
            "pattern": "^[^\\x00]*$",
            "title": "Password",
            "type": "string"
          },
          "phone": {
            "pattern": "^[^\\x00]*$",
            "title": "Phone",
            "type": "string"
          }
        },
        "required": [
          "first_name",
          "last_name",
          "phone",
          "password",
          "consent_granted",
          "consent_text_version"
        ],
        "title": "FarmerEnrollRequest",
        "type": "object"
      },
      "FarmerResponse": {
        "properties": {
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          },
          "phone": {
            "title": "Phone",
            "type": "string"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "id",
          "first_name",
          "last_name",
          "phone",
          "user_id"
        ],
        "title": "FarmerResponse",
        "type": "object"
      },
      "GrantedVia": {
        "description": "How a consent grant was captured (04 \u00a75.2).",
        "enum": [
          "onboarder",
          "farmer_self"
        ],
        "title": "GrantedVia",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RequestValidationErrorResponse": {
        "description": "FastAPI's own 422 body for a request that fails schema validation.\n\nA route that also raises a *domain* 422 (`consent_required`,\n`invalid_code`, ...) declares 422 itself, which replaces FastAPI's default\n422 schema. Without this second shape the declaration would then promise\nonly the envelope, and every malformed body would violate the contract\n(P3b-B: Schemathesis sent a lone NUL byte as a body and found exactly that).",
        "properties": {
          "detail": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "required": [
          "detail"
        ],
        "title": "RequestValidationErrorResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Honey provenance and traceability: hash-chained audit log, Merkle anchoring via OpenTimestamps, PostGIS geospatial records.",
    "title": "ApiChain v2 Backend",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v2/farmers": {
      "post": {
        "operationId": "enroll_farmer_v2_farmers_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FarmerEnrollRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Malformed request body"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Enroll Farmer",
        "tags": [
          "farmers"
        ]
      }
    },
    "/v2/farmers/{farmer_id}/consents": {
      "get": {
        "description": "The current state of every purpose. A read, so no audit row (the PII\n`data_access` rows are Phase 5).",
        "operationId": "farmer_consents_v2_farmers__farmer_id__consents_get",
        "parameters": [
          {
            "in": "path",
            "name": "farmer_id",
            "required": true,
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "title": "Farmer Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Malformed request body"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Farmer Consents",
        "tags": [
          "farmers"
        ]
      },
      "post": {
        "description": "Grant or withdraw one consent purpose (03 \u00a77, 11 \u00a75).\n\nThe ledger keeps every choice; the newest row decides. `data_processing`\ncannot be withdrawn here: withdrawing it means \"stop processing my data\",\nwhich is the Phase 5 deletion workflow, and recording a withdrawal nothing\nhonours would be worse than refusing it.",
        "operationId": "record_farmer_consent_v2_farmers__farmer_id__consents_post",
        "parameters": [
          {
            "in": "path",
            "name": "farmer_id",
            "required": true,
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "title": "Farmer Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentRecordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Malformed request body"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/RequestValidationErrorResponse"
                    }
                  ],
                  "title": "Response 422 Record Farmer Consent V2 Farmers  Farmer Id  Consents Post"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Record Farmer Consent",
        "tags": [
          "farmers"
        ]
      }
    },
...
  }
}